remove compilation warnings from icc
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@101052 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 67d9025e7d50ec2c6825283d964333d8f3cb7076

+++ Online/Online/OMAlib/algorithms/OMACheckHolesAndSpikes.cpp
@@ -35,7 +35,7 @@ void OMACheckHolesAndSpikes::exec(TH1 &Histo,
   if( warn_thresholds.size() <m_npars ||  alarm_thresholds.size() <m_npars 
       || input_pars.size() <m_ninput )
     return;
-  if (Histo.GetEntries() == 0)
+  if ( (int) Histo.GetEntries() == 0)
     return; // do nothing for empty histograms
   if (input_pars[0] < 0.) input_pars[0]= -99.;
   int ipoly= (int)(input_pars[0] + .1 );
